refactor(game): split 2076-line dashboard into focused files under 250 lines (#728)
Reduces app/game/page.tsx from 2076 lines to 65 lines by extracting:

- 18 presentation components into app/game/_components/dashboard/
  (HeroCard, RecommendedAction, LandsCard, ArmyCard, ThreatCard,
  ShieldCard, MiniMap, ExploreFrontier, BulkDistribute, BulkUnassign,
  DashboardHeader, DashboardReports, EligibilityBanner, NavGrid,
  CreatePlayerGate, NameYourGeneralGate, SignedOutLanding, DashboardView)

- 6 lib modules into app/game/_lib/
  (use-dashboard-data hook, dashboard-actions API handlers,
  dashboard-fetch initial-load, dashboard-helpers pure functions,
  dashboard-types shared types, recommend recommendation logic)

The _components and _lib prefixes are Next.js private folders, so the
new structure does not introduce any new routes. page.tsx is now a thin
gate that delegates to useDashboardData() and renders one of the gates
or the composed DashboardView.

Action handlers now take a DashboardMutators bag instead of 6+ closure
params each. Initial-load fetch logic is split from action handlers so
both files stay under 250 lines.

Adds unit tests for the extracted pure helpers (asDate, deriveShieldStatus,
describeShieldRemaining, countUnshieldedNeighbors, formatCountdown).

No behavior change. All 1600 tests pass; tsc clean; lint clean.
